surfaceflinger: allow targets to override the electronbeam duration
The Electron beam animation is constrained by the framerate the device
can achieve on GL. On adreno200, and even 205, the original 24 frames
resulted in a fairly smooth animation; on Tegra2, the animation is just
too fast.
From experimentation, I'd say the total number of frames should never be
less than half the attainable framerate (i.e., a half-second animation).

To use this, set TARGET_ELECTRONBEAM_FRAMES to half the desired value
(the default is 12 for a 24-frame animation)

Change-Id: I875b74ee3969ba0c8d38c234f7cb0f4402581ffb